Update server.go
As a beginner many, including myself, my use this hello world sample as a starting point to learn and test. However c.String can't be used for much more, at lest for a beginner, than to just do a simple test. Once I started to add lines the rendering was all off. It took me 2 days to find/figure out I had to use c.HTML instead of c.String to output HTML pages. So, perhaps, may be bet to have c.HTML in the hello world instead as to make it easier for newcomers to use the code as a starting point.