response: clarify Content-Disposition, close file in Stream example
It wasn't clear to me what were the differences between `Context#File`,
`Context#Attachment` and `Context#Inline` - reading the source made it
clearer: the former 2 are referring to sending `Content-Disposition`
with `attachment` and `inline`, which makes sense.

Adding missing `defer f.Close()` in the `Context#Stream` example. I also
checked, it takes an `io.Reader` not an `io.ReadCloser` so it couldn't
delegate the close to `Context#Stream`.